% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineScaling.R
\name{affineScaling}
\alias{affineScaling}
\title{Title}
\usage{
affineScaling(A, x, C)
}
\arguments{
\item{A}{The constraint Matrix}

\item{x}{initial values}

\item{C}{Objective Function Matrix}
}
\value{
Optimal solutions and objective Function value
}
\description{
Title
}
\examples{
affineScaling(A,x,C)
}
